Allow complete control of shouldIndex logic by DataObjects
The current extension point in the shouldIndex() method only
triggers after all default logic has been exhausted, limiting the
ability of DataObjects to define their own rules for indexing -
for example, to allow private content to enter the index.

This change allows DataObjects / extensions to directly replace
the logic of shouldIndex, and adds documentation warning of the
implications of ignoring the default logic.