Allow delegating control of shouldIndex logic to DataObjects (#21)
* Allow complete control of shouldIndex logic by DataObjects

The current extension point in the shouldIndex() method only
triggers after all default logic has been exhausted, limiting the
ability of DataObjects to define their own rules for indexing -
for example, to allow private content to enter the index.

This change allows DataObjects / extensions to directly replace
the logic of shouldIndex, and adds documentation warning of the
implications of ignoring the default logic.

* Add test coverage for new shouldIndex override

* Tweak override method name, throw on invalid result, tweak docs

* Fix new shouldIndex test coverage

* Shift to interface-based approach for shouldIndex override

* Tweak IndexableHandler interface name

* Update documentation to cover IndexableHandler

* Fix linting issues